requirements.txt
pillow
gradio
accelerate
datasets
peft
bitsandbytes
tensorboard